+++ test/index.spec.js
@@ -1,30 +1,30 @@
 import _ from 'lodash'
 import test from 'tape'
 import humps from '../'
+import { basicObj, basicRes } from './mock'
 
 test('humps', (t) => {
   const before = {
     'space key': 'space',
-    underscore_key: { another_one: 2 },
+    underscore_key: basicObj,
     TXT: 'ONE MORE THING',
     foo_bar: 'underscore',
     tests: 'ASTM D 4157-13',
-    contents: [ '100% Acrylic' ],
+    contents: ['100% Acrylic'],
     'dispatch-func': _.noop,
   }
   const after = {
     spaceKey: 'space',
-    underscoreKey: { anotherOne: 2 },
+    underscoreKey: basicRes,
     txt: 'ONE MORE THING',
     fooBar: 'underscore',
     tests: 'ASTM D 4157-13',
-    contents: [ '100% Acrylic' ],
-    dispatchFunc: {},
+    contents: ['100% Acrylic'],
+    dispatchFunc: _.noop,
   }
   t.equal(humps('some string_yo'), 'some string_yo', 'no string processing')
   t.deepEqual(humps(before), after, 'objects')
-  t.deepEqual(humps([ before, before ]), [ after, after ], 'array')
-  t.deepEqual(humps(_.noop), {}, 'kills functions')
-  t.deepEqual(humps(Date), {}, 'date class')
+  t.deepEqual(humps([before, before]), [after, after], 'array')
+  t.deepEqual(humps(_.noop), _.noop, 'saves functions')
   t.end()
 })