+++ test/mock.js
@@ -1,0 +1,2 @@
+export const basicObj = { attr_one: 'foo', attr_two: 'bar', attr_three: { attr_one: 'foo' } }
+export const basicRes = { attrOne: 'foo', attrTwo: 'bar', attrThree: { attrOne: 'foo' } }