+++ test/private.spec.js
@@ -1,12 +1,14 @@
 import test from 'tape'
-import { getVal, humpsObj } from '../src/lodash-humps'
+import { getVal, objIteratee } from '../src/lodash-humps'
+import { basicObj, basicRes } from './mock'
 
 test('getVal', (t) => {
   t.equal(getVal('foo bar'), 'foo bar', 'non object')
-  t.deepEqual(getVal({ foo_bar: 'under score' }), { fooBar: 'under score' }, 'object')
+  t.deepEqual(getVal(basicObj), basicRes, 'object')
   t.end()
 })
-test('humpsObj', (t) => {
-  t.deepEqual(humpsObj({ foo_bar: 'under score' }), { fooBar: 'under score' }, 'basic')
+test('objIteratee', (t) => {
+  t.deepEqual(objIteratee({}, 'under score', 'foo_bar'), { fooBar: 'under score' }, 'basic')
+  t.deepEqual(objIteratee({}, basicObj, 'baz'), { baz: basicRes })
   t.end()
 })